refactor to fix rubocop violations
