+++ lib/rsmp.rb
@@ -88,6 +88,12 @@ require_relative 'rsmp/tlc/modules/display'
 require_relative 'rsmp/tlc/modules/helpers'
 require_relative 'rsmp/tlc/startup_sequence'
 require_relative 'rsmp/tlc/traffic_controller_site'
+require_relative 'rsmp/tlc/proxy/control'
+require_relative 'rsmp/tlc/proxy/detectors'
+require_relative 'rsmp/tlc/proxy/io'
+require_relative 'rsmp/tlc/proxy/plans'
+require_relative 'rsmp/tlc/proxy/status'
+require_relative 'rsmp/tlc/proxy/system'
 require_relative 'rsmp/tlc/traffic_controller_proxy'
 require_relative 'rsmp/tlc/traffic_controller'
 require_relative 'rsmp/tlc/detector_logic'
