+++ spec/rsmp/supervisor/modules/configuration_spec.rb
